Revert "23 configuration 버그 수정 (#24)"
This reverts commit 2099d3cadff3362222343558669ecbddc97d1c01.

+++ srcs/config/child_config/server_config/include/IServerConfig.hpp
@@ -12,7 +12,7 @@ class IServerConfig : public IChildConfig {
  public:
   virtual void addLocationConfig(ILocationConfig* location_config) = 0;
   virtual std::list<ILocationConfig*> getLocationConfigs() = 0;
-  virtual short getListen() = 0;
+  virtual size_t getListen() = 0;
   const virtual std::string& getServerName() = 0;
   const virtual std::string& getErrorPage() = 0;
   const virtual std::string& getAccessLog() = 0;
